fix(spa): split /transactions/\$id into a layout + index so edit route renders
The detail route was a leaf with a child (edit), but its component
returned the detail UI directly without an <Outlet/>. Navigating to
/transactions/\$id/edit silently fell back to rendering the parent's
detail content instead of the matched child.

Move the detail UI into transactions.\$id.index.tsx (the default
child) and make transactions.\$id.tsx a pure layout that just
renders <Outlet/>. Same pattern already used for /transactions
(layout + index).
